requirements.txt
requests==2.14.2
python-ldap==2.4.38
PyYAML==3.12
